refactor altered imports and un-needed store imports
