+++ client/src/StreamComponents/fileDirectory/FileDirectory.jsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { Tree } from 'antd';
 import 'antd/dist/antd.css';
 import { connect } from 'react-redux';
-import { updateFile } from '../../redux/actions/index.js';
-import store from '../../redux/store/index.js';
+import { updateFile } from '../../redux/ducks/directoryDuck';
 
 const TreeNode = Tree.TreeNode;
